Rename "steps" folder
