+++ components/step-progress/step.tsx
