+++ components/step-progress/step1.tsx
