+++ components/step-progress/stepNavigation.tsx
